Fix integer overflow in Utilities.DisplayFileSizeToBytes
The method returns long but cast each result to int, so any value >= 2 GB
(and large mb inputs) overflowed to a negative/garbage number. Cast to long
instead and keep the gb multiplication in double space.

+++ src/libse/Common/Utilities.cs
@@ -111,7 +111,7 @@ namespace Nikse.SubtitleEdit.Core.Common
             {
                 if (double.TryParse(displayFileSize.Replace("bytes", string.Empty).Trim(), NumberStyles.AllowDecimalPoint, CultureInfo.CurrentCulture, out var n))
                 {
-                    return (int)Math.Round(n);
+                    return (long)Math.Round(n);
                 }
             }
 
@@ -119,7 +119,7 @@ namespace Nikse.SubtitleEdit.Core.Common
             {
                 if (double.TryParse(displayFileSize.Replace("kb", string.Empty).Trim(), NumberStyles.AllowDecimalPoint, CultureInfo.CurrentCulture, out var n))
                 {
-                    return (int)Math.Round(n * 1024);
+                    return (long)Math.Round(n * 1024);
                 }
             }
 
@@ -127,7 +127,7 @@ namespace Nikse.SubtitleEdit.Core.Common
             {
                 if (double.TryParse(displayFileSize.Replace("mb", string.Empty).Trim(), NumberStyles.AllowDecimalPoint, CultureInfo.CurrentCulture, out var n))
                 {
-                    return (int)Math.Round(n * 1024 * 1024);
+                    return (long)Math.Round(n * 1024 * 1024);
                 }
             }
 
@@ -135,7 +135,7 @@ namespace Nikse.SubtitleEdit.Core.Common
             {
                 if (double.TryParse(displayFileSize.Replace("gb", string.Empty).Trim(), NumberStyles.AllowDecimalPoint, CultureInfo.CurrentCulture, out var n))
                 {
-                    return (int)Math.Round(n * 1024 * 1024 * 1024);
+                    return (long)Math.Round(n * 1024.0 * 1024.0 * 1024.0);
                 }
             }
